doc_retrieval/requirements.txt
docling>=2.0,<3.0
httpx>=0.27,<1.0
faiss-cpu>=1.8,<2.0
numpy>=1.26,<3.0
tiktoken>=0.7,<1.0
openai>=1.30,<2.0
hydra-core>=1.3,<2.0
omegaconf>=2.3,<3.0
rank-bm25>=0.2,<1.0
